+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 2.0.0 - [Unreleased]
+
+* Rename `render` to `withRender` (breaking change).
+* Rename `window.*` to `window.hyperappRender.*` (breaking change).
+* Add support for
+  [lazy components](https://github.com/hyperapp/hyperapp/tree/d91e4667ee4e684eb874235e46ce919f502d4aae#lazy-components).
 
 ## [1.3.0] - 2018-02-24